add: routes/workout.js
Add routes to create and fetch workoutSessions from the database. It validates the parameters from the routes, payload and authentication.
It also validates the responses, to keep them standard through the routes.
If an error is thrown in returns 500 and console.log the error (for now), a better logger should be implemented in the future.