add: index.js
Main script of the application. It declares a webserver that will host the REST API. For the authentication method it uses JWT. And it creates a connection to mongodb as well to be used in the routes.
It dynamically registers the routes declares in the directory `routes`, same wise for the views.
It uses Swagger to auto generate API documentation.